fix CI problems
